pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "tgai-agent"
version = "1.0.0"
description = "Production-grade AI-powered Telegram agent platform with multi-provider AI, sub-agents, task scheduling, and plugin system"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "tgai-agent contributors" }]
keywords = ["telegram", "ai", "bot", "agent", "openai", "gemini", "claude", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Framework :: AsyncIO",
]
requires-python = ">=3.11"
dependencies = [
    "telethon>=1.36.0",
    "python-telegram-bot[job-queue]>=21.0",
    "openai>=1.30.0",
    "google-generativeai>=0.7.0",
    "anthropic>=0.28.0",
    "aiofiles>=23.2.0",
    "apscheduler>=3.10.4",
    "aiosqlite>=0.20.0",
    "cryptography>=42.0.0",
    "RestrictedPython>=7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "tenacity>=8.3.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "black>=24.4.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[project.urls]
Homepage = "https://github.com/your-username/tgai-agent"
Repository = "https://github.com/your-username/tgai-agent"
Documentation = "https://github.com/your-username/tgai-agent#readme"
"Bug Tracker" = "https://github.com/your-username/tgai-agent/issues"

[project.scripts]
tgai-agent = "tgai_agent.main:cli_entry"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"tgai_agent" = ["py.typed"]

# ── Black ─────────────────────────────────────────────────────────────────────
[tool.black]
line-length = 100
target-version = ["py311", "py312"]

# ── Ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "TID"]
ignore = ["E501", "B008", "TID252"]

[tool.ruff.lint.isort]
known-first-party = ["tgai_agent"]

# ── MyPy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
warn_unused_ignores = true
warn_return_any = false

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short --cov=src/tgai_agent --cov-report=term-missing --cov-report=xml"

[tool.coverage.run]
omit = ["tests/*", "*/migrations/*"]
